Extract request helper listener to separate class
