+++ data/src/main/kotlin/co/anitrend/arch/data/request/extension/Extensions.kt
@@ -17,23 +17,21 @@
 package co.anitrend.arch.data.request.extension
 
 import co.anitrend.arch.data.request.AbstractRequestHelper
-import co.anitrend.arch.data.request.contract.IRequestHelper
+import co.anitrend.arch.data.request.listener.RequestHelperListener
 import co.anitrend.arch.data.request.model.Request
 import co.anitrend.arch.data.request.report.contract.IRequestStatusReport
 import co.anitrend.arch.domain.entities.LoadState
 import co.anitrend.arch.domain.entities.RequestError
 import kotlinx.coroutines.channels.awaitClose
-import kotlinx.coroutines.channels.onFailure
 import kotlinx.coroutines.flow.callbackFlow
-import timber.log.Timber
 
-private fun IRequestStatusReport.getRequestError(): RequestError {
+internal fun IRequestStatusReport.getRequestError(): RequestError {
     return Request.Type.values().mapNotNull {
         getErrorFor(it)
     }.first()
 }
 
-private fun IRequestStatusReport.getTypeLoadPosition() = when (getType()) {
+internal fun IRequestStatusReport.getTypeLoadPosition() = when (getType()) {
     Request.Type.INITIAL -> LoadState.Position.UNDEFINED
     Request.Type.BEFORE -> LoadState.Position.TOP
     Request.Type.AFTER -> LoadState.Position.BOTTOM
@@ -43,28 +41,7 @@ private fun IRequestStatusReport.getTypeLoadPosition() = when (getType()) {
  * Creates a live data observable on the paging request helper
  */
 internal fun AbstractRequestHelper.createStatusFlow() = callbackFlow<LoadState> {
-    val requestListener = object : IRequestHelper.Listener {
-        /**
-         * Called when the status for any of the requests has changed.
-         *
-         * @param report The current status report that has all the information about the requests.
-         */
-        override fun onStatusChange(report: IRequestStatusReport) {
-            val position = report.getTypeLoadPosition()
-            val state = when {
-                report.hasRunning() -> LoadState.Loading(position)
-                report.hasError() -> {
-                    val error = report.getRequestError()
-                    LoadState.Error(error, position)
-                }
-                report.hasSuccess() -> LoadState.Success(position)
-                else -> LoadState.Idle(position)
-            }
-            trySend(state)
-                .onFailure { Timber.e(it) }
-        }
-    }
+    val requestListener = RequestHelperListener(scope = this)
     addListener(requestListener)
-
     awaitClose { removeListener(requestListener) }
 }
